Semi-working test cases
The "syncing adds post" does not work because the Twitter gem is not
accepting my fake Twitter response. Twitter replies with JSON, so I
have a suitable (I believe) response being passed back in the form of a
string. I tried doing a hash (like for the GitHub mock), but Webmock
does not accept hashes (Arrays are fine, though). This throws a syntax
error.

In other news, "syncing updates posts" works fine.